Tests/LibWeb/Text/input/DOM/Element-innerText-style-change-refreshes-segmenters.html

<!DOCTYPE html>
<script src="../include.js"></script>
<div id="target">a  b</div>
<script>
    test(() => {
        const target = document.getElementById("target");
        println(`initial: "${target.innerText}"`);

        target.style.textTransform = "uppercase";
        println(`after text-transform: "${target.innerText}"`);

        target.style.whiteSpace = "pre";
        println(`after white-space: "${target.innerText}"`);
    });
</script>
