Tests/LibWeb/Text/expected/DOM/Element-innerText-style-change-refreshes-segmenters.txt

initial: "a b"
after text-transform: "A B"
after white-space: "A  B"
